Put non-`RawEGraph` behind feature
